Move codegen parts of LoopStmt.cpp to codegen/
